requirements.txt
absl-py==0.13.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.6
anaconda-project==0.8.3
appdirs==1.4.4
asgiref==3.2.10
asn1crypto==0.24.0
astor==0.8.1
astroid==2.1.0
astropy==3.1
astunparse==1.6.3
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
bitarray @ file:///C:/ci/bitarray_1598885049384/work
bkcharts==0.2
blaze==0.11.3
bleach==3.0.2
blinker==1.4
bokeh==1.0.2
boto==2.49.0
boto3==1.12.0
botocore==1.15.0
Bottleneck==1.3.2
brotlipy==0.7.0
bz2file==0.98
cached-property==1.5.2
cachetools @ file:///tmp/build/80754af9/cachetools_1596822027882/work
certifi==2021.5.30
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
clyent==1.2.2
colorama==0.4.3
comtypes==1.1.7
conda==4.10.1
conda-build==3.17.6
conda-package-handling==1.7.0
conda-verify==3.1.1
contextlib2==0.5.5
cryptography==2.4.2
cycler==0.10.0
Cython==0.29.2
cytoolz==0.10.1
dask==1.0.0
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
distlib==0.3.1
distributed==1.25.1
Django==3.0.8
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.10
Flask==1.0.2
Flask-Cors==3.0.7
flatbuffers==1.12
fsspec @ file:///tmp/build/80754af9/fsspec_1597944003862/work
future==0.17.1
gast==0.4.0
gensim==3.8.0
gevent==1.3.7
glob2==0.6
google-api-core @ file:///C:/ci/google-api-core-split_1594356932141/work
google-auth @ file:///tmp/build/80754af9/google-auth_1599595709135/work
google-auth-oauthlib==0.4.1
google-cloud-core @ file:///tmp/build/80754af9/google-cloud-core_1599164813100/work
google-cloud-storage @ file:///tmp/build/80754af9/google-cloud-storage_1594357999796/work
google-pasta==0.2.0
google-resumable-media @ file:///tmp/build/80754af9/google-resumable-media_1594358346190/work
googleapis-common-protos==1.52.0
greenlet==0.4.15
grpcio==1.34.1
h5py==3.1.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.8
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==1.7.0
intel-openmp==2019.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.2
Jinja2==2.10
jmespath==0.9.4
joblib==0.14.1
json5==0.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core==4.6.3
jupyterlab==0.35.3
jupyterlab-server==0.2.0
Keras==2.4.3
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==17.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.9
llvmlite==0.26.0
locket==0.2.0
lxml==4.2.5
Markdown==3.1.1
MarkupSafe==1.1.0
matplot==0.1.9
matplotlib==3.3.1
mccabe==0.6.1
menuinst==1.4.16
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl==2019.0
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==3.0.5
more-itertools==4.3.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
mysqlclient==2.0.1
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose @ file:///C:/ci/nose_1594377973236/work
notebook==5.7.4
numba==0.41.0
numexpr==2.7.1
numpy==1.19.5
numpydoc==0.8.0
oauthlib==3.1.0
odo==0.5.1
olefile==0.46
opencv-python==4.5.1.48
openpyxl==2.5.12
opt-einsum==3.3.0
packaging==18.0
pandas @ file:///C:/ci/pandas_1598371525254/work
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pickleshare @ file:///C:/ci/pickleshare_1594374056827/work
Pillow==5.3.0
pipenv==2020.6.2
pkginfo==1.4.2
pluggy==0.8.0
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==2.0.10
protobuf==3.12.2
psutil==5.4.8
py==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.5
pyflakes==2.0.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.2.2
pyloco==0.0.139
pyodbc==4.0.25
pyOpenSSL==18.0.0
pyparsing==2.3.0
PyQt5==5.9.2
pyreadline==2.1
pyrsistent==0.14.11
PySocks==1.6.8
pytest==4.0.2
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.2.0
pytest-openfiles==0.3.1
pytest-remotedata==0.3.1
python-dateutil==2.7.5
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QtAwesome==0.7.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1598374667791/work
QtPy==1.9.0
readme-renderer==26.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rope==0.17.0
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
ruamel-yaml==0.15.87
s3transfer @ file:///C:/ci/s3transfer_1594411925509/work
scikit-image==0.16.2
scikit-learn @ file:///C:/ci/scikit-learn_1598376983131/work
scipy @ file:///C:/ci/scipy_1597686737426/work
seaborn==0.10.1
Send2Trash==1.5.0
simplegeneric==0.8.1
SimpleWebSocketServer==0.1.1
singledispatch==3.4.0.3
sip==4.19.8
six==1.15.0
sklearn==0.0
smart-open @ file:///tmp/build/80754af9/smart_open_1594236772807/work
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
Sphinx @ file:///tmp/build/80754af9/sphinx_1597428793432/work
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport @ file:///tmp/build/80754af9/sphinxcontrib-websupport_1597081412696/work
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy @ file:///C:/ci/sqlalchemy_1598374801163/work
sqlparse==0.3.1
statsmodels==0.11.1
sympy @ file:///C:/ci/sympy_1597083371142/work
tables==3.4.4
tblib @ file:///tmp/build/80754af9/tblib_1597928476713/work
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-hub==0.8.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toolz==0.10.0
tornado==5.1.1
tqdm @ file:///tmp/build/80754af9/tqdm_1596810128862/work
traitlets==4.3.3
twine==3.2.0
typing==3.7.4.3
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1597086586889/work
ushlex==0.99.1
virtualenv==20.0.30
virtualenv-clone==0.5.4
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.4.2
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter @ file:///tmp/build/80754af9/xlsxwriter_1598375848976/work
xlwings==0.20.5
xlwt==1.3.0
zict==2.0.0
zipp==3.1.0
